[Auth refactoring]: Migrate error handling to ServiceError pattern
- Add NoSessionTokenError, SessionTokenExpiredError, TokenRefreshError to ServiceError
- Migrate AuthResult to use ServiceError instead of AuthError
- Update AuthService to use ServiceError subclasses (UnexpectedError, NetworkError, StorageError)
- Update AuthProvider to throw NoSessionTokenError instead of NoSessionTokenFoundException
- Extract AuthState into separate file (auth_state.dart) with dartdoc
- Delete auth_error.dart and auth_exception.dart (replaced by ServiceError)
- Update unit tests to use new error types
- Add mocktail dependency for testing
