+++ pubspec.yaml
@@ -90,6 +90,7 @@ dev_dependencies:
   # riverpod_generator: ^2.1.3
   # riverpod_lint: ^2.3.9
   mockito: ^5.4.6
+  mocktail: ^1.0.4
   integration_test:
     sdk: flutter
   golden_toolkit: ^0.15.0
